+++ src/HIE/Bios/Cradle.hs
@@ -1,7 +1,8 @@
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections #-}
-{-# LANGUAGE BangPatterns #-}
 module HIE.Bios.Cradle (
       findCradle
     , loadCradle
@@ -22,48 +23,84 @@ module HIE.Bios.Cradle (
     , makeCradleResult
   ) where
 
+import Control.Applicative ((<|>), optional)
+import Data.Bifunctor (first)
+import Control.DeepSeq
 import Control.Exception (handleJust)
 import qualified Data.Yaml as Yaml
 import Data.Void
 import Data.Char (isSpace)
-import Data.Bifunctor (first)
-import System.Process
 import System.Exit
-import HIE.Bios.Types hiding (ActionName(..))
-import qualified HIE.Bios.Types as Types
-import HIE.Bios.Config
-import HIE.Bios.Environment (getCacheDir)
-import qualified HIE.Bios.Ghc.Gap as Gap
 import System.Directory hiding (findFile)
+import Control.Monad
+import Control.Monad.Extra (unlessM)
 import Control.Monad.Trans.Cont
 import Control.Monad.Trans.Maybe
-import System.FilePath
-import Control.Monad
-import System.Info.Extra
 import Control.Monad.IO.Class
-import System.Environment
-import Control.Applicative ((<|>), optional)
-import System.IO.Temp
-import System.IO.Error (isPermissionError)
-import Data.List
-import Data.Ord (Down(..))
-
-import System.PosixCompat.Files
-import HIE.Bios.Wrappers
-import System.IO
-import Control.DeepSeq
-
 import Data.Conduit.Process
 import qualified Data.Conduit.Combinators as C
 import qualified Data.Conduit as C
 import qualified Data.Conduit.Text as C
-import qualified Data.Text as T
 import qualified Data.HashMap.Strict as Map
-import           Data.Maybe (fromMaybe, maybeToList)
-import           GHC.Fingerprint (fingerprintString)
+import Data.Maybe (fromMaybe, maybeToList)
+import Data.List
+import Data.List.Extra (trimEnd)
+import Data.Ord (Down(..))
+import qualified Data.Text as T
+import System.Environment
+import System.FilePath
+import System.PosixCompat.Files
+import System.Info.Extra (isWindows)
+import System.IO (hClose, hGetContents, hSetBuffering, BufferMode(LineBuffering), withFile, IOMode(..))
+import System.IO.Error (isPermissionError)
+import System.IO.Temp
 
+import HIE.Bios.Config
+import HIE.Bios.Environment (getCacheDir)
+import HIE.Bios.Types hiding (ActionName(..))
+import HIE.Bios.Wrappers
+import qualified HIE.Bios.Types as Types
+import qualified HIE.Bios.Ghc.Gap as Gap
+
+import GHC.Fingerprint (fingerprintString)
+import GHC.ResponseFile (escapeArgs)
+
+----------------------------------------------------------------
+-- Environment variables used by hie-bios.
+--
+-- If you need more, add a constant here.
+----------------------------------------------------------------
+
+-- | Environment variable containing the filepath to which
+-- cradle actions write their results to.
+-- If the filepath does not exist, cradle actions must create them.
 hie_bios_output :: String
 hie_bios_output = "HIE_BIOS_OUTPUT"
+
+-- | Environment variable pointing to the GHC location used by
+-- cabal's and stack's GHC wrapper.
+--
+-- If not set, will default to sensible defaults.
+hie_bios_ghc :: String
+hie_bios_ghc = "HIE_BIOS_GHC"
+
+-- | Environment variable with extra arguments passed to the GHC location
+-- in cabal's and stack's GHC wrapper.
+--
+-- If not set, assume no extra arguments.
+hie_bios_ghc_args :: String
+hie_bios_ghc_args = "HIE_BIOS_GHC_ARGS"
+
+-- | Environment variable pointing to the source file location that caused
+-- the cradle action to be executed.
+hie_bios_arg :: String
+hie_bios_arg = "HIE_BIOS_ARG"
+
+-- | Environment variable pointing to a filepath to which dependencies
+-- of a cradle can be written to by the cradle action.
+hie_bios_deps :: String
+hie_bios_deps = "HIE_BIOS_DEPS"
+
 ----------------------------------------------------------------
 
 -- | Given root\/foo\/bar.hs, return root\/hie.yaml, or wherever the yaml file was found.
@@ -98,7 +135,7 @@ getCradle :: (b -> Cradle a) -> (CradleConfig b, FilePath) -> Cradle a
 getCradle buildCustomCradle (cc, wdir) = addCradleDeps cradleDeps $ case cradleType cc of
     Cabal CabalType{ cabalComponent = mc } -> cabalCradle wdir mc
     CabalMulti dc ms ->
-      getCradle buildCustomCradle $
+      getCradle buildCustomCradle
         (CradleConfig cradleDeps
           (Multi [(p, CradleConfig [] (Cabal $ dc <> c)) | (p, c) <- ms])
         , wdir)
@@ -108,7 +145,7 @@ getCradle buildCustomCradle (cc, wdir) = addCradleDeps cradleDeps $ case cradleT
       in
         stackCradle wdir mc stackYamlConfig
     StackMulti ds ms ->
-      getCradle buildCustomCradle $
+      getCradle buildCustomCradle
         (CradleConfig cradleDeps
           (Multi [(p, CradleConfig [] (Stack $ ds <> c)) | (p, c) <- ms])
         , wdir)
@@ -129,14 +166,12 @@ addCradleDeps deps c =
     addActionDeps :: CradleAction a -> CradleAction a
     addActionDeps ca =
       ca { runCradle = \l fp ->
-            runCradle ca l fp
-              >>= \case
-                CradleSuccess (ComponentOptions os' dir ds) ->
-                  pure $ CradleSuccess (ComponentOptions os' dir (ds `union` deps))
-                CradleFail err ->
-                  pure $ CradleFail
-                    (err { cradleErrorDependencies = cradleErrorDependencies err `union` deps })
-                CradleNone -> pure CradleNone
+            (cradleLoadResult
+                CradleNone
+                (\err -> CradleFail (err { cradleErrorDependencies = cradleErrorDependencies err `union` deps }))
+                (\(ComponentOptions os' dir ds) -> CradleSuccess (ComponentOptions os' dir (ds `union` deps)))
+            )
+            <$> runCradle ca l fp
          }
 
 -- | Try to infer an appropriate implicit cradle type from stuff we can find in the enclosing directories:
@@ -334,7 +369,7 @@ multiAction buildCustomCradle cur_dir cs l cur_fp =
     canonicalizeCradles :: IO [(FilePath, CradleConfig b)]
     canonicalizeCradles =
       sortOn (Down . fst)
-        <$> mapM (\(p, c) -> (,c) <$> (makeAbsolute (cur_dir </> p))) cs
+        <$> mapM (\(p, c) -> (,c) <$> makeAbsolute (cur_dir </> p)) cs
 
     selectCradle [] =
       return (CradleFail (CradleError [] ExitSuccess err_msg))
@@ -398,7 +433,7 @@ biosAction :: FilePath
 biosAction wdir bios bios_deps l fp = do
   bios' <- callableToProcess bios (Just fp)
   (ex, _stdo, std, [(_, res),(_, mb_deps)]) <-
-    readProcessWithOutputs [hie_bios_output, "HIE_BIOS_DEPS"] l wdir bios'
+    readProcessWithOutputs [hie_bios_output, hie_bios_deps] l wdir bios'
 
   deps <- case mb_deps of
     Just x  -> return x
@@ -412,9 +447,7 @@ biosAction wdir bios bios_deps l fp = do
 callableToProcess :: Callable -> Maybe String -> IO CreateProcess
 callableToProcess (Command shellCommand) file = do
   old_env <- getEnvironment
-  return $ (shell shellCommand) { env = (: old_env) <$> (,) hieBiosArg <$> file }
-    where
-      hieBiosArg = "HIE_BIOS_ARG"
+  return $ (shell shellCommand) { env = (: old_env) . (,) hie_bios_arg <$> file }
 callableToProcess (Program path) file = do
   canon_path <- canonicalizePath path
   return $ proc canon_path (maybeToList file)
@@ -428,18 +461,117 @@ cabalCradle wdir mc =
     { cradleRootDir    = wdir
     , cradleOptsProg   = CradleAction
         { actionName = Types.Cabal
-        , runCradle = cabalAction wdir mc
-        , runGhcCmd = \args -> do
-            buildDir <- cabalBuildDir wdir
+        , runCradle = \l f -> runCradleResultT . cabalAction wdir mc l $ f
+        , runGhcCmd = \args -> runCradleResultT $ do
+            buildDir <- liftIO $ cabalBuildDir wdir
             -- Workaround for a cabal-install bug on 3.0.0.0:
             -- ./dist-newstyle/tmp/environment.-24811: createDirectory: does not exist (No such file or directory)
-            createDirectoryIfMissing True (buildDir </> "tmp")
+            liftIO $ createDirectoryIfMissing True (buildDir </> "tmp")
             -- Need to pass -v0 otherwise we get "resolving dependencies..."
-            wrapper_fp <- withCabalWrapperTool ("ghc", []) wdir
-            readProcessWithCwd
-              wdir "cabal" (["--builddir="<>buildDir,"v2-exec","--with-compiler", wrapper_fp, "ghc", "-v0", "--"] ++ args) ""
+            cabalProc <- cabalProcess wdir "v2-exec" $ ["ghc", "-v0", "--"] ++ args
+            readProcessWithCwd' cabalProc ""
         }
     }
+
+-- | Execute a cabal process in our custom cache-build directory configured
+-- with the custom ghc executable.
+-- The created process has its working directory set to the given working directory.
+--
+-- Invokes the cabal process in the given directory.
+-- Finds the appropriate @ghc@ version as a fallback and provides the path
+-- to the custom ghc wrapper via 'hie_bios_ghc' environment variable which
+-- the custom ghc wrapper may use as a fallback if it can not respond to certain
+-- queries, such as ghc version or location of the libdir.
+cabalProcess :: FilePath -> String -> [String] -> CradleLoadResultT IO CreateProcess
+cabalProcess workDir command args = do
+  ghcDirs <- cabalGhcDirs workDir
+  newEnvironment <- liftIO $ setupEnvironment ghcDirs
+  cabalProc <- liftIO $ setupCabalCommand ghcDirs
+  pure $ (cabalProc
+      { env = Just newEnvironment
+      , cwd = Just workDir
+      })
+  where
+    processEnvironment :: (FilePath, FilePath) -> [(String, String)]
+    processEnvironment (ghcBin, libdir) =
+      [(hie_bios_ghc, ghcBin), (hie_bios_ghc_args,  "-B" ++ libdir)]
+
+    setupEnvironment :: (FilePath, FilePath) -> IO [(String, String)]
+    setupEnvironment ghcDirs = do
+      environment <- getCleanEnvironment
+      pure $ processEnvironment ghcDirs ++ environment
+
+    setupCabalCommand :: (FilePath, FilePath) -> IO CreateProcess
+    setupCabalCommand (ghcBin, libdir) = do
+      wrapper_fp <- withGhcWrapperTool ("ghc", []) workDir
+      buildDir <- cabalBuildDir workDir
+      ghcPkgPath <- withGhcPkgTool ghcBin libdir
+      let extraCabalArgs =
+            [ "--builddir=" <> buildDir
+            , command
+            , "--with-compiler", wrapper_fp
+            , "--with-hc-pkg", ghcPkgPath
+            ]
+      pure $ proc "cabal" (extraCabalArgs ++ args)
+
+-- | Discovers the location of 'ghc-pkg' given the absolute path to 'ghc'
+-- and its '$libdir' (obtainable by running @ghc --print-libdir@).
+--
+-- @'withGhcPkgTool' ghcPathAbs libdir@ guesses the location by looking at
+-- the filename of 'ghcPathAbs' and expects that 'ghc-pkg' is right next to it,
+-- which is guaranteed by the ghc build system. Most OS's follow this
+-- convention.
+--
+-- On unix, there is a high-chance that the obtained 'ghc' location is the
+-- "unwrapped" executable, e.g. the executable without a shim that specifies
+-- the '$libdir' and other important constants.
+-- As such, the executable 'ghc-pkg' is similarly without a wrapper shim and
+-- is lacking certain constants such as 'global-package-db'. It is, therefore,
+-- not suitable to pass in to other consumers, such as 'cabal'.
+--
+-- Here, we restore the wrapper-shims, if necessary, thus the returned filepath
+-- can be passed to 'cabal' without further modifications.
+withGhcPkgTool :: FilePath -> FilePath -> IO FilePath
+withGhcPkgTool ghcPathAbs libdir = do
+  let ghcName = takeFileName ghcPathAbs
+      -- TODO: check for existence
+      ghcPkgPath = guessGhcPkgFromGhc ghcName
+  if isWindows
+    then pure ghcPkgPath
+    else withWrapperTool ghcPkgPath
+  where
+    ghcDir = takeDirectory ghcPathAbs
+
+    guessGhcPkgFromGhc ghcName =
+      let ghcPkgName = T.replace "ghc" "ghc-pkg" (T.pack ghcName)
+      in ghcDir </> T.unpack ghcPkgName
+
+    -- Only on unix, creates a wrapper script that's hopefully identical
+    -- to the wrapper script 'ghc-pkg' usually comes with.
+    --
+    -- 'ghc-pkg' needs to know the 'global-package-db' location which is
+    -- passed in via a wrapper shim that basically wraps 'ghc-pkg' and
+    -- only passes in the correct 'global-package-db'.
+    -- For an example on how the wrapper script is supposed to look like, take
+    -- a look at @cat $(which ghc-pkg)@, assuming 'ghc-pkg' is on your $PATH.
+    --
+    -- If we used the raw executable, i.e. not wrapped in a shim, then 'cabal'
+    -- can not use the given 'ghc-pkg'.
+    withWrapperTool ghcPkg = do
+      let globalPackageDb = libdir </> "package.conf.d"
+          -- This is the same as the wrapper-shims ghc-pkg usually comes with.
+          contents = unlines
+            [ "#!/bin/sh"
+            , unwords ["exec", escapeFilePath ghcPkg
+                      , "--global-package-db", escapeFilePath globalPackageDb
+                      , "${1+\"$@\"}"
+                      ]
+            ]
+          srcHash = show (fingerprintString contents)
+      cacheFile "ghc-pkg" srcHash $ \wrapperFp -> writeFile wrapperFp contents
+
+    -- Escape the filepath and trim excess newlines added by 'escapeArgs'
+    escapeFilePath fp = trimEnd $ escapeArgs [fp]
 
 -- | @'cabalCradleDependencies' rootDir componentDir@.
 -- Compute the dependencies of the cabal cradle based
@@ -473,8 +605,7 @@ processCabalWrapperArgs args =
             let final_args =
                     removeVerbosityOpts
                     $ removeRTS
-                    $ removeInteractive
-                    $ ghc_args
+                    $ removeInteractive ghc_args
             in Just (dir, final_args)
         _ -> Nothing
 
@@ -483,69 +614,112 @@ processCabalWrapperArgs args =
 -- arguments to the executable.
 type GhcProc = (FilePath, [String])
 
--- | Generate a fake GHC that can be passed to cabal
+-- | Generate a fake GHC that can be passed to cabal or stack
 -- when run with --interactive, it will print out its
 -- command-line arguments and exit
-withCabalWrapperTool :: GhcProc -> FilePath -> IO FilePath
-withCabalWrapperTool (mbGhc, ghcArgs) wdir = do
-    cacheDir <- getCacheDir ""
-    createDirectoryIfMissing True cacheDir
+withGhcWrapperTool :: GhcProc -> FilePath -> IO FilePath
+withGhcWrapperTool (mbGhc, ghcArgs) wdir = do
     let wrapperContents = if isWindows then cabalWrapperHs else cabalWrapper
-        suffix fp = if isWindows then fp <.> "exe" else fp
-    let srcHash = show (fingerprintString wrapperContents)
-    let wrapper_name = "wrapper-" ++ srcHash
-    let wrapper_fp = suffix $ cacheDir </> wrapper_name
-    exists <- doesFileExist wrapper_fp
-    unless exists $
+        withExtension fp = if isWindows then fp <.> "exe" else fp
+        srcHash = show (fingerprintString wrapperContents)
+    cacheFile (withExtension "wrapper") srcHash $ \wrapper_fp ->
       if isWindows
-      then do
+      then
         withSystemTempDirectory "hie-bios" $ \ tmpDir -> do
-          let wrapper_hs = cacheDir </> wrapper_name <.> "hs"
+          let wrapper_hs = wrapper_fp -<.> "hs"
           writeFile wrapper_hs wrapperContents
           let ghc = (proc mbGhc $
                       ghcArgs ++ ["-rtsopts=ignore", "-outputdir", tmpDir, "-o", wrapper_fp, wrapper_hs])
                       { cwd = Just wdir }
           readCreateProcess ghc "" >>= putStr
       else writeFile wrapper_fp wrapperContents
-    setMode wrapper_fp
-    pure wrapper_fp
+
+-- | Create and cache a file in hie-bios's cache directory.
+--
+-- @'cacheFile' fpName srcHash populate@. 'fpName' is the pattern name of the
+-- cached file you want to create. 'srcHash' is the hash that is appended to
+-- the file pattern and is expected to change whenever you want to invalidate
+-- the cache.
+--
+-- If the cached file's 'srcHash' changes, then a new file is created, but
+-- the old cached file name will not be deleted.
+--
+-- If the file does not exist yet, 'populate' is invoked with cached file
+-- location and it is expected that the caller persists the given filepath in
+-- the File System.
+cacheFile :: FilePath -> String -> (FilePath -> IO ()) -> IO FilePath
+cacheFile fpName srcHash populate = do
+  cacheDir <- getCacheDir ""
+  createDirectoryIfMissing True cacheDir
+  let newFpName = cacheDir </> (dropExtensions fpName <> "-" <> srcHash) <.> takeExtensions fpName
+  unlessM (doesFileExist newFpName) $ do
+    populate newFpName
+    setMode newFpName
+  pure newFpName
   where
     setMode wrapper_fp = setFileMode wrapper_fp accessModes
 
 -- | Given the root directory, get the build dir we are using for cabal
 -- In the `hie-bios` cache directory
 cabalBuildDir :: FilePath -> IO FilePath
-cabalBuildDir work_dir = do
-  abs_work_dir <- makeAbsolute work_dir
+cabalBuildDir workDir = do
+  abs_work_dir <- makeAbsolute workDir
   let dirHash = show (fingerprintString abs_work_dir)
-  getCacheDir ("dist-"<>filter (not . isSpace) (takeBaseName abs_work_dir)<>"-"<>dirHash)
+  getCacheDir ("dist-" <> filter (not . isSpace) (takeBaseName abs_work_dir)<>"-"<>dirHash)
 
-cabalAction :: FilePath -> Maybe String -> LoggingFunction -> FilePath -> IO (CradleLoadResult ComponentOptions)
-cabalAction work_dir mc l fp = do
-    wrapper_fp <- withCabalWrapperTool ("ghc", []) work_dir
-    buildDir <- cabalBuildDir work_dir
-    let cab_args = ["--builddir="<>buildDir,"v2-repl", "--with-compiler", wrapper_fp, fromMaybe (fixTargetPath fp) mc]
-    (ex, output, stde, [(_,mb_args)]) <-
-      readProcessWithOutputs [hie_bios_output] l work_dir (proc "cabal" cab_args)
-    let args = fromMaybe [] mb_args
-    case processCabalWrapperArgs args of
-        Nothing -> do
-          -- Best effort. Assume the working directory is the
-          -- the root of the component, so we are right in trivial cases at least.
-          deps <- cabalCradleDependencies work_dir work_dir
-          pure $ CradleFail (CradleError deps ex
-                    ["Failed to parse result of calling cabal"
-                     , unlines output
-                     , unlines stde
-                     , unlines $ args])
-        Just (componentDir, final_args) -> do
-          deps <- cabalCradleDependencies work_dir componentDir
-          pure $ makeCradleResult (ex, stde, componentDir, final_args) deps
+-- | Discover the location of the ghc binary 'cabal' is going to use together
+-- with its libdir location.
+-- The ghc executable is an absolute path, but not necessarily canonicalised
+-- or normalised. Additionally, the ghc path returned is likely to be the raw
+-- executable, i.e. without the usual wrapper shims on non-windows systems.
+-- If you want to use the given ghc executable, you should invoke
+-- 'withGhcWrapperTool'.
+--
+-- If cabal can not figure it out, a 'CradleError' is returned.
+cabalGhcDirs :: FilePath -> CradleLoadResultT IO (FilePath, FilePath)
+cabalGhcDirs workDir = do
+  libdir <- readProcessWithCwd_ workDir "cabal" ["exec", "-v0", "--", "ghc", "--print-libdir"] ""
+  exe <- readProcessWithCwd_ workDir "cabal"
+      -- DON'T TOUCH THIS CODE
+      -- This works with 'NoImplicitPrelude', with 'RebindableSyntax' and other shenanigans.
+      -- @-package-env=-@ doesn't work with ghc prior 8.4.x
+      [ "exec", "-v0", "--" , "ghc", "-package-env=-", "-ignore-dot-ghci", "-e"
+      , "Control.Monad.join (Control.Monad.fmap System.IO.putStr System.Environment.getExecutablePath)"
+      ]
+      ""
+  pure (trimEnd exe, trimEnd libdir)
+
+cabalAction :: FilePath -> Maybe String -> LoggingFunction -> FilePath -> CradleLoadResultT IO ComponentOptions
+cabalAction workDir mc l fp = do
+  cabalProc <- cabalProcess workDir "v2-repl" [fromMaybe (fixTargetPath fp) mc] `modCradleError` \err -> do
+      deps <- cabalCradleDependencies workDir workDir
+      pure $ err { cradleErrorDependencies = cradleErrorDependencies err ++ deps }
+
+  (ex, output, stde, [(_, maybeArgs)]) <- liftIO $ readProcessWithOutputs [hie_bios_output] l workDir cabalProc
+
+  let args = fromMaybe [] maybeArgs
+  case processCabalWrapperArgs args of
+    Nothing -> do
+      -- Provide some dependencies an IDE can look for to trigger a reload.
+      -- Best effort. Assume the working directory is the
+      -- root of the component, so we are right in trivial cases at least.
+      deps <- liftIO $ cabalCradleDependencies workDir workDir
+      throwCE (CradleError deps ex $
+                [ "Failed to parse result of calling cabal"
+                , "Failed command: " <> prettyCmdSpec (cmdspec cabalProc)
+                , unlines output
+                , unlines stde
+                , unlines $ args
+                , "Process Environment:"]
+                <> prettyProcessEnv cabalProc)
+    Just (componentDir, final_args) -> do
+      deps <- liftIO $ cabalCradleDependencies workDir componentDir
+      CradleLoadResultT $ pure $ makeCradleResult (ex, stde, componentDir, final_args) deps
   where
     -- Need to make relative on Windows, due to a Cabal bug with how it
     -- parses file targets with a C: drive in it
     fixTargetPath x
-      | isWindows && hasDrive x = makeRelative work_dir x
+      | isWindows && hasDrive x = makeRelative workDir x
       | otherwise = x
 
 removeInteractive :: [String] -> [String]
@@ -621,14 +795,13 @@ stackCradle wdir mc syaml =
     , cradleOptsProg   = CradleAction
         { actionName = Types.Stack
         , runCradle = stackAction wdir mc syaml
-        , runGhcCmd = \args -> do
+        , runGhcCmd = \args -> runCradleResultT $ do
             -- Setup stack silently, since stack might print stuff to stdout in some cases (e.g. on Win)
             -- Issue 242 from HLS: https://github.com/haskell/haskell-language-server/issues/242
-            stackSetup <- readProcessWithCwd wdir "stack" (stackYamlProcessArgs syaml <> ["setup", "--silent"]) ""
-            stackSetup `bindIO` \_ ->
-              readProcessWithCwd wdir "stack"
-                (stackYamlProcessArgs syaml <> ["exec", "ghc", "--"] <> args)
-                ""
+            _ <- readProcessWithCwd_ wdir "stack" (stackYamlProcessArgs syaml <> ["setup", "--silent"]) ""
+            readProcessWithCwd_ wdir "stack"
+              (stackYamlProcessArgs syaml <> ["exec", "ghc", "--"] <> args)
+              ""
         }
     }
 
@@ -650,26 +823,26 @@ stackCradleDependencies wdir componentDir syaml = do
     cabalFiles ++ [relFp </> "package.yaml", stackYamlLocationOrDefault syaml]
 
 stackAction :: FilePath -> Maybe String -> StackYaml -> LoggingFunction -> FilePath -> IO (CradleLoadResult ComponentOptions)
-stackAction work_dir mc syaml l _fp = do
+stackAction workDir mc syaml l _fp = do
   let ghcProcArgs = ("stack", stackYamlProcessArgs syaml <> ["exec", "ghc", "--"])
   -- Same wrapper works as with cabal
-  wrapper_fp <- withCabalWrapperTool ghcProcArgs work_dir
-  (ex1, _stdo, stde, [(_, mb_args)]) <-
-    readProcessWithOutputs [hie_bios_output] l work_dir $
+  wrapper_fp <- withGhcWrapperTool ghcProcArgs workDir
+  (ex1, _stdo, stde, [(_, maybeArgs)]) <-
+    readProcessWithOutputs [hie_bios_output] l workDir $
     stackProcess syaml
                 $  ["repl", "--no-nix-pure", "--with-ghc", wrapper_fp]
                     <> [ comp | Just comp <- [mc] ]
   (ex2, pkg_args, stdr, _) <-
-    readProcessWithOutputs [hie_bios_output] l work_dir $
+    readProcessWithOutputs [hie_bios_output] l workDir $
       stackProcess syaml ["path", "--ghc-package-path"]
   let split_pkgs = concatMap splitSearchPath pkg_args
       pkg_ghc_args = concatMap (\p -> ["-package-db", p] ) split_pkgs
-      args = fromMaybe [] mb_args
+      args = fromMaybe [] maybeArgs
   case processCabalWrapperArgs args of
       Nothing -> do
         -- Best effort. Assume the working directory is the
         -- the root of the component, so we are right in trivial cases at least.
-        deps <- stackCradleDependencies work_dir work_dir syaml
+        deps <- stackCradleDependencies workDir workDir syaml
         pure $ CradleFail
                   (CradleError deps ex1 $
                     [ "Failed to parse result of calling stack" ]
@@ -678,7 +851,7 @@ stackAction work_dir mc syaml l _fp = do
                   )
 
       Just (componentDir, ghc_args) -> do
-        deps <- stackCradleDependencies work_dir componentDir syaml
+        deps <- stackCradleDependencies workDir componentDir syaml
         pure $ makeCradleResult
                   ( combineExitCodes [ex1, ex2]
                   , stde ++ stdr, componentDir
@@ -730,15 +903,15 @@ bazelCommand :: String
 bazelCommand = $(embedStringFile "wrappers/bazel")
 
 rulesHaskellAction :: FilePath -> FilePath -> IO (CradleLoadResult ComponentOptions)
-rulesHaskellAction work_dir fp = do
+rulesHaskellAction workDir fp = do
   wrapper_fp <- writeSystemTempFile "wrapper" bazelCommand
   setFileMode wrapper_fp accessModes
-  let rel_path = makeRelative work_dir fp
+  let rel_path = makeRelative workDir fp
   (ex, args, stde) <-
-      readProcessWithOutputFile work_dir wrapper_fp [rel_path] []
+      readProcessWithOutputFile workDir wrapper_fp [rel_path] []
   let args'  = filter (/= '\'') args
   let args'' = filter (/= "\"$GHCI_LOCATION\"") (words args')
-  deps <- rulesHaskellCradleDependencies work_dir
+  deps <- rulesHaskellCradleDependencies workDir
   return $ makeCradleResult (ex, stde, args'') deps
 
 
@@ -769,11 +942,11 @@ obeliskCradle wdir =
     }
 
 obeliskAction :: FilePath -> FilePath -> IO (CradleLoadResult ComponentOptions)
-obeliskAction work_dir _fp = do
+obeliskAction workDir _fp = do
   (ex, args, stde) <-
-      readProcessWithOutputFile work_dir "ob" ["ide-args"] []
+      readProcessWithOutputFile workDir "ob" ["ide-args"] []
 
-  o_deps <- obeliskCradleDependencies work_dir
+  o_deps <- obeliskCradleDependencies workDir
   return (makeCradleResult (ex, stde, words args) o_deps )
 
 -}
@@ -808,13 +981,12 @@ findFile p dir = do
     getFiles = filter p <$> getDirectoryContents dir
     doesPredFileExist file = doesFileExist $ dir </> file
 
--- Some environments (e.g. stack exec) include GHC_PACKAGE_PATH.
--- Cabal v2 *will* complain, even though or precisely because it ignores them
+-- | Some environments (e.g. stack exec) include GHC_PACKAGE_PATH.
+-- Cabal v2 *will* complain, even though or precisely because it ignores them.
 -- Unset them from the environment to sidestep this
 getCleanEnvironment :: IO [(String, String)]
 getCleanEnvironment = do
-  e <- getEnvironment
-  return $ Map.toList $ Map.delete "GHC_PACKAGE_PATH" $ Map.fromList e
+  Map.toList . Map.delete "GHC_PACKAGE_PATH" . Map.fromList <$> getEnvironment
 
 type Outputs = [OutputName]
 type OutputName = String
@@ -831,16 +1003,17 @@ readProcessWithOutputs
   -> FilePath -- ^ Working directory. Process is executed in this directory.
   -> CreateProcess -- ^ Parameters for the process to be executed.
   -> IO (ExitCode, [String], [String], [(OutputName, Maybe [String])])
-readProcessWithOutputs outputNames l work_dir cp = flip runContT return $ do
+readProcessWithOutputs outputNames l workDir cp = flip runContT return $ do
   old_env <- liftIO getCleanEnvironment
   output_files <- traverse (withOutput old_env) outputNames
 
   let process = cp { env = Just $ output_files ++ fromMaybe old_env (env cp),
-                     cwd = Just work_dir
+                     cwd = Just workDir
                     }
 
     -- Windows line endings are not converted so you have to filter out `'r` characters
-  let  loggingConduit = C.decodeUtf8  C..| C.lines C..| C.filterE (/= '\r')  C..| C.map T.unpack C..| C.iterM l C..| C.sinkList
+  let loggingConduit = C.decodeUtf8  C..| C.lines C..| C.filterE (/= '\r')
+        C..| C.map T.unpack C..| C.iterM l C..| C.sinkList
   (ex, stdo, stde) <- liftIO $ sourceProcessWithStreams process mempty loggingConduit loggingConduit
 
   res <- forM output_files $ \(name,path) ->
@@ -890,15 +1063,46 @@ runGhcCmdOnPath wdir args = readProcessWithCwd wdir "ghc" args ""
   -- case mResult of
   --   Nothing
 
--- | Wrapper around 'readCreateProcess' that sets the working directory
+-- | Wrapper around 'readCreateProcess' that sets the working directory and
+-- clears the environment, suitable for invoking cabal/stack and raw ghc commands.
 readProcessWithCwd :: FilePath -> FilePath -> [String] -> String -> IO (CradleLoadResult String)
-readProcessWithCwd dir cmd args stdi = do
-  cleanEnv <- getCleanEnvironment
+readProcessWithCwd dir cmd args stdin = runCradleResultT $ readProcessWithCwd_ dir cmd args stdin
+
+readProcessWithCwd_ :: FilePath -> FilePath -> [String] -> String -> CradleLoadResultT IO String
+readProcessWithCwd_ dir cmd args stdin = do
+  cleanEnv <- liftIO getCleanEnvironment
   let createProc = (proc cmd args) { cwd = Just dir, env = Just cleanEnv }
-  mResult <- optional $ readCreateProcessWithExitCode createProc stdi
+  readProcessWithCwd' createProc stdin
+
+-- | Wrapper around 'readCreateProcessWithExitCode', wrapping the result in
+-- a 'CradleLoadResult'. Provides better error messages than raw 'readCreateProcess'.
+readProcessWithCwd' :: CreateProcess -> String -> CradleLoadResultT IO String
+readProcessWithCwd' createdProcess stdin = do
+  mResult <- liftIO $ optional $ readCreateProcessWithExitCode createdProcess stdin
+  let cmdString = prettyCmdSpec $ cmdspec createdProcess
   case mResult of
-    Just (ExitSuccess, stdo, _) -> pure $ CradleSuccess stdo
-    Just (exitCode, stdo, stde) -> pure $ CradleFail $
-      CradleError [] exitCode ["Error when calling " <> cmd <> " " <> unwords args, stdo, stde]
-    Nothing -> pure $ CradleFail $
-      CradleError [] ExitSuccess ["Couldn't execute " <> cmd <> " " <> unwords args]
+    Just (ExitSuccess, stdo, _) -> pure stdo
+    Just (exitCode, stdo, stde) -> throwCE $
+      CradleError [] exitCode $
+        ["Error when calling " <> cmdString, stdo, stde] <> prettyProcessEnv createdProcess
+    Nothing -> throwCE $
+      CradleError [] ExitSuccess $
+        ["Couldn't execute " <> cmdString] <> prettyProcessEnv createdProcess
+
+-- | Prettify 'CmdSpec', so we can show the command to a user
+prettyCmdSpec :: CmdSpec -> String
+prettyCmdSpec (ShellCommand s) = s
+prettyCmdSpec (RawCommand cmd args) = cmd ++ " " ++ unwords args
+
+-- | Pretty print hie-bios's relevant environment variables.
+prettyProcessEnv :: CreateProcess -> [String]
+prettyProcessEnv p =
+  [ key <> ": " <> value
+  | (key, value) <- fromMaybe [] (env p)
+  , key `elem` [ hie_bios_output
+               , hie_bios_ghc
+               , hie_bios_ghc_args
+               , hie_bios_arg
+               , hie_bios_deps
+               ]
+  ]